move kitchen sink to games, split into modules, camera to update
- Moved from samples/features/feature-everything to samples/games/game-everything
- Split 1473-line main.rs into 6 modules: state, title, game, pause, countdown, main
- Moved camera setup (follow, dead zone, bounds, shake, zoom, rotation) from render() to update()
- Moved clear_color from render() to update()
- Removed Cell<bool> hack for pending_shake — plain bool works now that camera lives in update()
- TitleScene and CountdownScene draw entirely in update() (no render needed)
- GameScene keeps render() for scene-stack overlay support (pause menu)
- PauseOverlay keeps render() for same reason
- Renamed package from rengine-feature-everything to rengine-everything
- Updated ARCHITECTURE.md and workspace Cargo.toml
- Headless integration test passes (all 32+ features verified)
