+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["engine", "samples/games/game-platformer", "samples/games/game-topdown", "samples/games/game-iso", "samples/games/game-fps", "samples/games/game-fight", "samples/games/game-fps-mp", "samples/features/feature-scenes", "samples/features/feature-sprites", "samples/features/feature-camera", "samples/features/feature-input", "samples/features/feature-animation", "samples/features/feature-canvas", "samples/features/feature-pixelart", "samples/features/feature-tilemap", "samples/features/feature-triggers", "samples/features/feature-everything", "samples/features/feature-rng", "samples/features/feature-nineslice", "samples/features/feature-tween", "samples/features/feature-text", "samples/features/feature-ui", "samples/features/feature-saveload", "samples/features/feature-resolution", "samples/features/feature-particles", "samples/features/feature-audio", "samples/features/feature-postfx"]
+members = ["engine", "samples/games/game-platformer", "samples/games/game-topdown", "samples/games/game-iso", "samples/games/game-fps", "samples/games/game-fight", "samples/games/game-fps-mp", "samples/games/game-everything", "samples/features/feature-scenes", "samples/features/feature-sprites", "samples/features/feature-camera", "samples/features/feature-input", "samples/features/feature-animation", "samples/features/feature-canvas", "samples/features/feature-pixelart", "samples/features/feature-tilemap", "samples/features/feature-triggers", "samples/features/feature-rng", "samples/features/feature-nineslice", "samples/features/feature-tween", "samples/features/feature-text", "samples/features/feature-ui", "samples/features/feature-saveload", "samples/features/feature-resolution", "samples/features/feature-particles", "samples/features/feature-audio", "samples/features/feature-postfx"]
 resolver = "2"
