+++ samples/games/game-everything/assets/coin_sfx.wav
