+++ samples/games/game-everything/assets/game_config.json
